Searching while typing
Whenever the end-user enters text into `<input type="search">` element,
an [input][] event fill fire fore each keystroke and bubble up.. When
the `<input>` element's text changes, we'll refresh the search results
by submitting the corresponding `<form>` element. Since the `<form>`
element targets the `<turbo-frame id="search_results">`, the frame will
navigate automatically whenever the form submits.

To submit the `<form>` after each keystroke, we'll declare an action
descriptor to route all `input` events to the `form#submit` action:

```diff
--- a/app/views/layouts/application.html.erb
+++ b/app/views/layouts/application.html.erb
@@ -17,11 +17,11 @@
   <body>
     <header>
       <form action="<%= searches_path(turbo_frame: "search_results") %>" data-turbo-frame="search_results" class="peer"
-        data-controller="form" data-action="invalid->form#hideValidationMessage:capture">
+        data-controller="form" data-action="invalid->form#hideValidationMessage:capture input->form#submit">
         <label for="search_query">Query</label>
         <input id="search_query" name="query" type="search" pattern=".*\w+.*" required>

         <button>
           Search
         </button>
       </form>
```

We'll declare `[data-form-target="submit"]` on the `<form>` element's
`<button>`, so that the `form` controller instance can access it as a
[Target][]:

```diff
--- a/app/views/layouts/application.html.erb
+++ b/app/views/layouts/application.html.erb
@@ -17,11 +17,11 @@
-        <button>
+        <button data-form-target="submit">
           Search
         </button>
```

We'll declare the `submit` target within the `form` controller, then
implement the `submit()` action to reference the new target and submit
the `<form>` by clicking the `<button>` on the end-user's behalf:

```diff
--- a/app/javascript/controllers/form_controller.js
+++ b/app/javascript/controllers/form_controller.js
 import { Controller } from "@hotwired/stimulus"

 export default class extends Controller {
+  static get targets() { return [ "submit" ] }
+
+  submit() {
+    this.submitTarget.click()
+  }
+
   hideValidationMessage(event) {
     event.stopPropagation()
     event.preventDefault()
   }
 }
```

Since we'll be automatically submitting the form on each keystroke, we
have an opportunity to hide the submit button. We'll use JavaScript to
set the [hidden][] attribute on the element. By deferring the `[hidden]`
attribute to JavaScript, we can ensure that the element is visible
whenever end-users are browsing without JavaScript enabled.

```diff
--- a/app/javascript/controllers/form_controller.js
+++ b/app/javascript/controllers/form_controller.js
 import { Controller } from "@hotwired/stimulus"

 export default class extends Controller {
   static get targets() { return [ "submit" ] }
+
+  connect() {
+    this.submitTarget.hidden = true
+  }

   submit() {
```

Finally, submitting the `<form>` on _every_ keystroke results in a
cascade of sequential HTTP requests. In that scenario, the majority of
the intermediate responses could be ignored. To limit the number of
concurrent requests, we'll add limitations to ensure that a submission
occurs _at most_ once every 200 milliseconds. We'll add a client-side
dependency on [Lodash.debounce][] through [Skypack][]. Once that
function is available, we'll re-bind the `form#submit` action to the
debounced version:

```diff
--- a/app/javascript/controllers/form_controller.js
+++ b/app/javascript/controllers/form_controller.js
 import { Controller } from "@hotwired/stimulus"
+import debounce from "https://cdn.skypack.dev/lodash.debounce"

 export default class extends Controller {
   static get targets() { return [ "submit" ] }
+
+  initialize() {
+    this.submit = debounce(this.submit.bind(this), 200)
+  }

   connect() {
     this.submitTarget.hidden = true
   }
```

[input]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/input_event
[Target]: https://stimulus.hotwire.dev/handbook/building-something-real#defining-the-target
[hidden]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/hidden
[Lodash.debounce]: https://lodash.com/docs/4.17.15#debounce
[Skypack]: https://www.skypack.dev/about

Wrapping up
---

We started with an out-of-the-box Rails application generated by `rails
new`, then implemented a collapsible search-as-you-type experience. The
`<input type="search">` element's container expands to show its results
in-line while searching, supports keyboard navigation and selection, and
only submits requests to the server when a search term is present.

We never encode our search request or response into JSON, and our client
communicates with our server without any calls to [XMLHttpRequest][] or
[fetch][] from within our application code.

On top of all that, we've implemented the experience with [semantically
meaningful][] elements like [`<form>`][mdn-form], [`<input
type="search">`][mdn-input-search], and [`<mark>`][mdn-mark]!

[semantically meaningful]: https://developer.mozilla.org/en-US/docs/Glossary/semantics
[XMLHttpRequest]: https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest
[fetch]: https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API
[mdn-form]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form
[mdn-input-search]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/search
[mdn-mark]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/mark